feat: add a paper container
It adds a container to paper and moves the info and library components out of the paper component to avoid rendering them unnecessarily.
